Issue 230
https://github.com/HospitalRun/hospitalrun-frontend/issues/230#issuecomment-163484990

~\app\patients\edit\controller.js
Updated "afterUpdate" action to not use displayAlert. The display alert
was a helper function that helped users quickly open a message dialog.
Instead the "afterUpdate" action will now call a new dialog directly and
bypass the displayAlert helper class. The new updated afterUpdate action
specifies that the update button will call an action to bring the user
back to the patient list. The action that the updateButtonAction is set
to is called 'returnToPatient', this action has been added to the
controller file and the route file.

~\app\patients\edit\route.js
Added new action 'returnToPatient' points to the action in the
controller file.

~\app\templates\components\modal-dialog.hbs
Updated the model dialogue template to allow users to customize the
cancel button. In order to customize the cancel button you must set the
flag {{isCancelTextChanged}} to true and then set  {{cancelButtonText}}
to the text that you want to be displayed instead. If you do not have
the flag set to true, then the template will render the default cancel
button that has the text "Cancel" The reason for setting a flag is so
that other dialogs that use this template can continue to work normally
without having to go back to modify and set {{canceButtonText}} for
every instance of the template.

~\app\components\modal-dialog.js
Updated component to capture the custom text and flag if they wish to
customize the text

~\app\dialog\template.hbs
Updated the mapping to capture the user input.