Refactor patients/delete/controller to use e-c
Example of using ember-concurrency tasks to manage a complex promise
chain of events. I picked this one mainly because I wanted to illustrate
the simplicity of using e-c tasks to manage what was otherwise a very
complex promise chain.

I tried to preserve some of the concurrency described in the previous
promise based code. However, after some analysis and discussion on the
Ember Slack channels difference between preserving the concurrency and
just running the resolutions serially are likely very small. In this
case the use of `all()` could likely be removed without a significant
impact on performance. I mention this later optimisation as a way to
make the code even easier to grok. I'll leave the debate to further
discussion.