Rename deprecated jquery helpers to their own files
Some tests require on the jquery query engine to select elements
based on selectors that aren't compatible with document.querySelector
which the new ember test helpers use. To maintain this for now we
introduce a deprecated selector to warn against using these in the future,
but allow tests that currently need them to still pass.