Add link to v1.6.1
Due to the issues surrounding CouchDB v2.3.0, it would be very helpful to add a direct link to the archived location of v1.6.1 and mention that it is the recommended, working version of CouchDB to use with the HospitalRun project. The instructions currently direct you to download CouchDB, and then reveal, several steps later, that v2.3.0 (the version linked to) is experimental and not fully supported.

I spent a while scratching my head last night trying to clone and set up the project, and after I found CouchDB v1.6.1 on the Apache archives, everything immediately clicked and built just fine. Finding the particular page and version of CouchDB is a bit of a nested process, and linking directly to it, along with a note for the purpose of doing that, will be beneficial to anyone coming to HospitalRun for the first time.