test(diagnoses): added additional tests for hooks, routing, and loading
New tests were added to test the behavior of the useDiagnosis hook, the routing
of the onRowClick event handler, and the loading component for viewing a
diagnosis. Some code was also refactored to simplify the logic of the rendering.
Lastly, a bug that prevented the Diagnoses tab from being highlighted when
viewing a diagnosis was also fixed.